% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{draw_timeline}
\alias{draw_timeline}
\title{Helper function for geom_timeline}
\usage{
draw_timeline(data, panel_params, coord, na.rm = FALSE)
}
\arguments{
\item{data}{Dataframe. A dataframe with at least x, y and group aesthetics}

\item{panel_params}{List. Contains information about the scales in the current panel}

\item{coord}{Environment. Coordinate specification}

\item{na.rm}{boolean. Not implemented}
}
\value{
A grob to draw in the panel

grob points. A grob object with the desired layer
}
\description{
Internal usage.
}
\references{
\url{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}
}
