% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{draw_timeline_label}
\alias{draw_timeline_label}
\title{Helper function for geom_timeline_label base class}
\usage{
draw_timeline_label(data, panel_params, coord, check_overlap)
}
\arguments{
\item{data}{data.frame. A dataframe with at least x, y and group aesthetics}

\item{panel_params}{list. Contains information about the scales in the current panel}

\item{coord}{environment. Coordinate specification}

\item{check_overlap}{boolean. Omit overlaped text.}
}
\value{
grobTree. The text labels and corresponding guide lines.
}
\description{
Internal usage.
}
