% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_coordinates}
\alias{eq_clean_coordinates}
\title{Clean coordinates in the NOAA dataset}
\usage{
eq_clean_coordinates(dataframe)
}
\arguments{
\item{dataframe}{A tibble with the NOAA dataset}
}
\value{
A tibble
}
\description{
Coherce coordinates to numeric and removes non valid rows from the
NOAA raw dataset
}
\examples{
\dontrun{eq_get_data() \%>\% eq_clean_coordinates}

}
