% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans the NOAA dataset}
\usage{
eq_clean_data(dataframe)
}
\arguments{
\item{dataframe}{A tibble. The raw NOAA dataset}
}
\value{
A tibble
}
\description{
A simple wrapper for the main cleaning functions. This function will
clean dates, coordintes and location names.
}
\examples{
eq_get_data() \%>\% eq_clean_data
}
