% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_dates}
\alias{eq_clean_dates}
\title{Cleaning NOAA dates}
\usage{
eq_clean_dates(dataframe)
}
\arguments{
\item{dataframe}{A tibble with the NOAA raw dataframe}
}
\value{
tibble. A data frame with the parsed dates and removed BCE dates
}
\description{
The BCE dates where removed from the dataset because R cannot handle
this dates
}
\examples{
\dontrun{eq_get_data() \%>\% eq_clean_dates}

}
