% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_location}
\alias{eq_clean_location}
\title{Clean names for the location in the NOAA dataset}
\usage{
eq_clean_location(dataframe)
}
\arguments{
\item{dataframe}{A tibble. The NOAA raw dataset}
}
\value{
A tibble
}
\description{
This function removes the country name and converts location to titlecase
for later usage in labels
}
\examples{
\dontrun{eq_get_data() \%>\% eq_clean_location()}

}
