% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_wrong_parsed_numeric_vectors}
\alias{eq_clean_wrong_parsed_numeric_vectors}
\title{Clean magnitude variables}
\usage{
eq_clean_wrong_parsed_numeric_vectors(dataframe)
}
\arguments{
\item{dataframe}{A tibble. The raw NOAA dataset}
}
\value{
tibble. A data frame with magnitudes cleaned
}
\description{
Magnitude variables are parsed as character vectors. This functions mutate
them into numeric.
}
\examples{
\dontrun{
  eq_get_data() \%>\% eq_clean_worng_parsed_numeric_vectors
}

}
