% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafleat.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create annotation HTML text for leaflet}
\usage{
eq_create_label(dataframe)
}
\arguments{
\item{dataframe}{data.frame. The data to create labels.}
}
\value{
character. A custom HTML template
}
\description{
This function helps to create a custon HTML template for to annotate the leaflet.
The data to show is:
- Location
- Magnitude
- Total number of deaths
}
\examples{
library(lubridate)
library(dplyr)

eq_get_data() \%>\%
  eq_clean_data \%>\%
  filter(COUNTRY == "MEXICO", year(DATE) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
