% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{eq_download_data}
\alias{eq_download_data}
\title{Download the NOAA dataset}
\usage{
eq_download_data(destdir)
}
\arguments{
\item{destdir}{character. Destination directory}
}
\value{
NULL. Impure function, called by its side effects
}
\description{
Get the NOAA dataset in TXT file format. If the destination directory
don't exist, will be created.
}
\examples{
\dontrun{download_noaa()}

}
