% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{eq_get_data}
\alias{eq_get_data}
\title{Get NOAA dataframe}
\usage{
eq_get_data(destdir = NULL, update = FALSE)
}
\arguments{
\item{destdir}{character. A character vector with path to the working directory.
If NULL a persistet folder will be created (default and prefered).}

\item{update}{boolean. Delete the current file and download the dataset again?}
}
\value{
tibble. A tibble containing the full NOAA dataset
}
\description{
Securely give you the NOAA dataset. If you do not have the dataset it will
download the data in a persistent folder and give you the full data.frame.
The column type is character by default and coherent transformations can be
performed with the cleaning data functions of this same package.
}
\examples{
eq_get_data()
eq_get_data \%>\% class

}
