% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafleat.R
\name{eq_map}
\alias{eq_map}
\title{Draws a map to locate the epicenters}
\usage{
eq_map(dataframe, annot_col = NULL)
}
\arguments{
\item{dataframe}{data.frame. The filtered dataset containing the points to draw}

\item{annot_col}{character. A column name to create the popup labels.}
}
\value{
leaflet. A leaflet map
}
\description{
This function will create a leaflet map with points in the earthquakes
epicenters. Also you can pass a \code{annot_col} param to create a popup
in each marker
}
\examples{
library(dplyr)
library(lubridate)

eq_get_data() \%>\%
  eq_clean_data \%>\%
  filter(COUNTRY == "MEXICO", year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")

}
