% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Base class for geom_timeline_label}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomTimelineLabel
}
\description{
Base class for geom_timeline_label
}
\keyword{datasets}
