% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAACoursera.R
\docType{package}
\name{NOAACoursera-package}
\alias{NOAACoursera}
\alias{NOAACoursera-package}
\title{NOAACoursera - Coursera Capstone Project}
\description{
This is the final project of the Coursera's "Master Software Development in R"
specialization. The objective of this package is provide wrapper functions to
quickly analyse and visualize the National Oceanic and Atmospheric Administration
dataset of significant earthquakes.
}
