% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{numeric_convertible}
\alias{numeric_convertible}
\title{Tests if a vector can be safely converted to a numeric type}
\usage{
numeric_convertible(test_vector)
}
\arguments{
\item{test_vector}{character. A vector to test}
}
\value{
Boolean. TRUE if the vector can be safely converted to numeric.
}
\description{
This test will not take into account NA_character_ in the test. Only character
values will be tested.
}
