% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{StatTimelineLabel}
\alias{StatTimelineLabel}
\title{Base class for stat_timelinelabel}
\format{An object of class \code{StatTimelineLabel} (inherits from \code{Stat}, \code{ggproto}) of length 2.}
\usage{
StatTimelineLabel
}
\description{
Internal usage. This class computes the labels to be drawn in the
geom_timeline_label.
}
\keyword{datasets}
