% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_timeline.R
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{NOAA timeline ggplot theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 7.}
\usage{
theme_timeline
}
\description{
A simple theme fot the NOAACourser package
}
\keyword{datasets}
